Was interested in using an LRU cache with max age when I came across this repo. Was hoping to make a few changes/additions if interested! This contains major breaking changes.

Though we haven't been tagging releases, this could correspond to a minor bump or major bump to v2.0 if we'd previously considered the API stable. Afaik, this is only used in two repos, both of which use govendor to pin the sha.

Changes:
```
* Added additional stats around cache capacity, count, sets, gets,
  and evictions
* Sub was renamed to Delta as not all stats are counters
* New now accepts a config object with optional fields
* New panics on invalid configuration
* Size was renamed to Capacity
* MaxAge is optional, and a duration of 0 disables expiration
* Configuration accepts two new optional callbacks: OnEviction and
  OnExpiration. The eviction callback is only invoked on LRU policy
  eviction, and not on element removal or expiration as seen in
  some other pkgs
* Add was renamed to Set since we overwrite any previous value
* Added new funcs: Has, Peek, Remove, Len, Clear, Keys, OrderedKeys,
  and SetMageAge
* Reduced allocations
```

```
// master
$ make bench
go test --bench=. --benchmem
BenchmarkCache-8   	 3000000	       431 ns/op	      80 B/op	       3 allocs/op
PASS
ok  	github.com/segmentio/agecache	1.751s

// PR
go test --bench=. --benchmem
BenchmarkCache-8   	 3000000	       405 ns/op	      32 B/op	       2 allocs/op
PASS
ok  	github.com/segmentio/agecache	1.661s
```

cc @segmentio/gateway