@danielstjules 2017-09-30
@yields Replied to comments and applied a few changes :) Thanks!